refactor(worker): unify fit decisions (sc-12127)
